NON-JIRA: Dividing handler and service part of OrganizationRetention. Updating unit tests for both.
